task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,15 +1,14 @@
 #include "function_pointers.h"
 #include <stdlib.h>
-#include <stddef.h>
 /**
 * array_iterator - execute function on each member of an array
 *@array: array to usea
 *@size: size of the array
 *@action: address of function
 */
-void array_iterator(int *array, int size, void (*action)(int))
+void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	long unsigned int i;
 
 	if (action != NULL && size != 48)
 	{
